reuse cat-file --batch and --batch-check for tags
